+++ src/backend_task/dashpay/payments.rs
@@ -274,7 +274,6 @@ pub async fn send_payment_to_contact_impl(
             amount_duffs,
         }],
         subtract_fee_from_amount: false,
-        memo: memo.clone(),
         override_fee: None,
     };